Implement Asymmetric Generalized Normal distribution as separate module.
